+++ tests/review_helpers/cli.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 
 import unittest
 
-from l2tdevtools.helpers import cli as cli_helper
+from l2tdevtools.review_helpers import cli
 
 
 class CLIHelperTest(unittest.TestCase):
@@ -15,7 +15,7 @@ class CLIHelperTest(unittest.TestCase):
   def testRunCommand(self):
     """Tests that the helper can be initialized."""
     mock_responses = {'echo hi': [0, b'hi\n', b'']}
-    test_helper = cli_helper.CLIHelper(mock_responses=mock_responses)
+    test_helper = cli.CLIHelper(mock_responses=mock_responses)
     with self.assertRaises(AttributeError):
       test_helper.RunCommand('echo hello')
     exit_code, stdout, stderr = test_helper.RunCommand('echo hi')
@@ -23,7 +23,7 @@ class CLIHelperTest(unittest.TestCase):
     self.assertEqual(stdout, b'hi\n')
     self.assertEqual(stderr, b'')
 
-    real_helper = cli_helper.CLIHelper()
+    real_helper = cli.CLIHelper()
     exit_code, stdout, stderr = real_helper.RunCommand('echo hello')
     self.assertEqual(exit_code, 0)
     self.assertEqual(stdout, b'hello\n')
