+++ tests/review_helpers/projects.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 
 import unittest
 
-import l2tdevtools.helpers.projects as projects_helper
+from l2tdevtools.review_helpers import projects
 
 
 class ProjectsHelperTest(unittest.TestCase):
@@ -16,7 +16,7 @@ class ProjectsHelperTest(unittest.TestCase):
 
   def testInitialize(self):
     """Tests that the helper can be initialized."""
-    helper = projects_helper.ProjectsHelper('/home/plaso/l2tdevtools/review.py')
+    helper = projects.ProjectsHelper('/home/plaso/l2tdevtools/review.py')
     self.assertIsNotNone(helper)
 
   # TODO: add test for version_file_path property
@@ -27,16 +27,16 @@ class ProjectsHelperTest(unittest.TestCase):
     """Tests the GetReviewer function."""
     author = 'test@example.com'
 
-    reviewer = projects_helper.ProjectsHelper.GetReviewer('l2tdevtools', author)
+    reviewer = projects.ProjectsHelper.GetReviewer('l2tdevtools', author)
     self.assertNotEqual(reviewer, author)
-    self.assertIn(reviewer, projects_helper.ProjectsHelper._REVIEWERS_DEFAULT)
+    self.assertIn(reviewer, projects.ProjectsHelper._REVIEWERS_DEFAULT)
 
   def testGetReviewersOnCC(self):
     """Tests the GetReviewersOnCC function."""
     author = 'test@example.com'
     reviewer = 'joachim.metz@gmail.com'
 
-    reviewers_cc = projects_helper.ProjectsHelper.GetReviewersOnCC(
+    reviewers_cc = projects.ProjectsHelper.GetReviewersOnCC(
         'l2tdevtools', author, reviewer)
     self.assertNotIn(author, reviewers_cc)
     self.assertNotIn(reviewer, reviewers_cc)
